refactor: login, signup response, request type 위치 수정
