+++ src/types/api/signup/index.ts
@@ -1,8 +1,9 @@
-import type { LoginResponseType } from '../login';
+import { LoginResponseType } from '@/types/api/login';
+
 export interface SignupRequestType {
   email: string;
   fullName: string;
   password: string;
 }
 
-export interface SignupType extends LoginResponseType {}
+export interface SignupResponseType extends LoginResponseType {}
